Massive functional improvements and refactoring.
We added support for POST transactions. We fixed a number of
mutex issues. We added support for /api/query calls, including
special handling of format=json calls.

We will consider NOT caching /api/query results.

Some additional future plans would be to carry the expiration
time down to the individual cache entries. We would, when a
cache hit happens on a freed item check to see if the result
is the same as it was last time and we would bump up the lifetime
on that item. Perhaps we can do that with a hash value such
as an md5 (or even a CRC32?). The goal here is to try to
"tune" the lifetime of stable cache entries to reduce load
on the underlying servers as much as possible.

Jus thoughts right now.

Probably we will branch soon and start working on that
as the next major feature branch.